refactor: download file
