+++ src/error.rs
@@ -2,6 +2,7 @@ pub mod cli;
 pub mod dfx;
 pub mod dfxvm;
 pub mod dfxvm_init;
+pub mod download;
 pub mod env;
 pub mod fs;
 pub mod installation;
