+++ src/error/dfxvm/install.rs
@@ -1,12 +1,8 @@
 use crate::error::{
+    download::DownloadVerifiedTarballError,
     env::NoHomeDirectoryError,
-    fs::{
-        CreateDirAllError, CreateFileError, OpenFileError, ReadToStringError, RenameError,
-        WriteFileError,
-    },
+    fs::{CreateDirAllError, OpenFileError, RenameError},
     json::LoadJsonFileError,
-    reqwest::WrappedReqwestError,
-    Retryable,
 };
 use std::path::PathBuf;
 use thiserror::Error;
@@ -42,55 +38,6 @@ pub enum InstallError {
 }
 
 #[derive(Error, Debug)]
-pub enum DownloadVerifiedTarballError {
-    #[error(transparent)]
-    DownloadFile(#[from] DownloadFileError),
-
-    #[error("no such version")]
-    NoSuchVersion(#[source] WrappedReqwestError),
-
-    #[error("failed to parse url")]
-    ParseUrl(#[from] url::ParseError),
-
-    #[error(transparent)]
-    VerifyChecksum(#[from] VerifyChecksumError),
-}
-
-#[derive(Error, Debug)]
-pub enum DownloadFileError {
-    #[error(transparent)]
-    CreateFile(#[from] CreateFileError),
-
-    #[error("failed to download contents of {url}")]
-    DownloadContents {
-        url: String,
-        source: WrappedReqwestError,
-    },
-
-    #[error(transparent)]
-    Get(WrappedReqwestError),
-
-    #[error("failed to get content length from {url}")]
-    GetContentLength { url: String },
-
-    #[error(transparent)]
-    Status(WrappedReqwestError),
-
-    #[error(transparent)]
-    WriteFile(#[from] WriteFileError),
-}
-
-impl Retryable for DownloadFileError {
-    fn is_retryable(&self) -> bool {
-        match self {
-            DownloadFileError::DownloadContents { .. } => true,
-            DownloadFileError::Get(e) => e.is_retryable(),
-            _ => false,
-        }
-    }
-}
-
-#[derive(Error, Debug)]
 pub enum ExtractArchiveError {
     #[error(transparent)]
     OpenFile(#[from] OpenFileError),
@@ -100,16 +47,4 @@ pub enum ExtractArchiveError {
         path: PathBuf,
         source: std::io::Error,
     },
-}
-
-#[derive(Error, Debug)]
-pub enum VerifyChecksumError {
-    #[error("checksum did not match.  Expected={expected} Actual={actual}")]
-    HashMismatch { expected: String, actual: String },
-
-    #[error("checksum file is malformed. Expected the first word of the first line to contain a hash.  Actual contents: {contents}")]
-    MalformedChecksumFile { contents: String },
-
-    #[error(transparent)]
-    ReadToString(#[from] ReadToStringError),
 }
